process updates go in separate thread, ui decoupled from logic updates, pager fork initialized early on and initiates pager executions without additional dtrace overhead

+++ src/fiveman_process_state.c
@@ -1,20 +1,21 @@
 #include <assert.h>
 #include <errno.h>
+#include <fcntl.h>
 #include <signal.h>
+#include <spawn.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
 
 #include "fiveman.h"
 #include "fiveman_instruction.h"
+#include "fiveman_pager_fork.h"
 #include "fiveman_process_intent.h"
 #include "fiveman_process_state.h"
 #include "fiveman_process_state_table.h"
 #include "fiveman_process_statistics.h"
 #include "ncurses_screen.h"
 #include "signal_handlers.h"
-
-fiveman_process_state * state_in_fork = NULL;
 
 fiveman_process_state * fiveman_process_state_allocate(fiveman_instruction * instr, int port){
   fiveman_process_state * state = calloc(1, sizeof(fiveman_process_state));
@@ -23,16 +24,15 @@ fiveman_process_state * fiveman_process_state_allocate(fiveman_instruction * ins
   state->intent.intent = INTENT_NONE;
   state->desired_port = port;
   state->instruction = instr;
-  state->host_read_fd = -1;
-  state->child_write_fd = -1;
+  state->sample_ctxt = NULL;
   return state;
 }
 
 void fiveman_process_state_deallocate(fiveman_process_state * state){
   fiveman_process_state * next_ptr = NULL;
   while(state != NULL){
-    fiveman_process_state_close_pipes(1, 1, state);
     next_ptr = state->next;
+    fiveman_teardown_sampling(state);
     unlink(state->stdout);
     unlink(state->stderr);
     free(state->stdout);
@@ -71,9 +71,7 @@ void fiveman_process_state_initialize(fiveman_process_state * state){
 
 
 pid_t fiveman_process_state_start(fiveman_process_state * state, char * directory){
-  usleep((int) ((float)MICROSECONDS_PER_SECOND * 0.05));
   fiveman_process_state_clear_sample(state);
-  state_in_fork = NULL;
 
   const char * existing_path = getenv("PATH");
   const char * path_assignment = "PATH=";
@@ -81,73 +79,51 @@ pid_t fiveman_process_state_start(fiveman_process_state * state, char * director
   char new_path_buf[new_path_buf_size];
   snprintf(new_path_buf, new_path_buf_size, "%s%s:%s", path_assignment, directory, existing_path);
 
-  const char * port_assignment = "PORT=";
+    const char * port_assignment = "PORT=";
   const size_t new_port_buf_size = strlen(port_assignment) + 128;
   char new_port_buf[new_port_buf_size];
   snprintf(new_port_buf, new_port_buf_size, "%s%i", port_assignment, state->desired_port);
 
-  fiveman_process_state_initialize_pipes(state);
-  ignore_sigpipe();
-  pid_t child = fork();
-  if(child == 0){ // in communicating child
-    fiveman_process_state_table_close_pipes(0, 1, application_state_table);
-    assert(freopen(state->stderr, "a", stderr) != NULL);
-    assert(freopen(state->stdout, "a", stdout) != NULL);
-    assert(freopen("/dev/null", "r", stdin) != NULL);
-    restore_sigpipe();
-    state_in_fork = state;
-    assert(setpgid(0, 0) == 0);
-    pid_t subchild = fiveman_sampling_fork();
-    if(subchild == 0){ // in executing child
-      fiveman_process_state_close_pipes(1, 1, state);
-      assert(setgid(getgid()) == 0);
-      assert(setuid(getuid()) == 0);
-      assert(putenv(new_path_buf) == 0);
-      assert(putenv(new_port_buf) == 0);
-      assert(chdir(directory) == 0);
-      execl("/bin/sh", "/bin/sh", "-c", state->instruction->exec, NULL);
-      // If execution proceeds here, we should print error message and exit immediately
-      printf("Couldn't execute instruction %s, %i\n", state->instruction->exec, errno);
-      exit(errno);
-    } else if (subchild > 0){
-      // HACKY: The state now reflects the actual executing subchild here, as we are further sub-delegating management
-      // in this fork.
-      state_in_fork->pid = subchild;
-      int status = 0;
-      install_child_kill_handler();
-      fiveman_init_sampling(subchild);
-      time_t cur;
-      time(&cur);
-      fflush(stderr);
-      while(1){
-        waitpid(subchild, &status, WNOHANG);
-        if(!fiveman_process_state_is_alive(state)){
-          break;
-        }
-        if(state != NULL){
-          fiveman_process_state_sample_process(state);
-          write(state->child_write_fd, &state->sample, sizeof(fiveman_process_statistics_sample));
-        }
-        fiveman_sampling_sleep(cur);
-        fflush(stderr);
-        time(&cur);
-      }
-      waitpid(subchild, &status, 0);
-      fiveman_teardown_sampling(subchild);
-      exit(status);
-    } else {
-      assert(subchild >= 0);
-      exit(-1);
-    }
-  } else if(child > 0){ // in parent
-    restore_sigpipe();
-    fiveman_process_state_close_pipes(1, 0, state);
-    assert(setpgid(child, child) == 0);
-    return child;
-  } else {
-    assert(child >= 0);
-    exit(-1);
+  char * argv[] = { "/bin/sh", "-c", (char *) state->instruction->exec, NULL };
+
+  int esize = 0;
+  char * c = environ[0];
+  for( int i = 0; c != NULL; c = environ[i++]){
+    esize = i;
   }
+  char ** new_environ = calloc(esize + 2, sizeof(char *));
+  memcpy(new_environ, environ, esize * sizeof(char *));
+  new_environ[esize - 1] = new_path_buf;
+  new_environ[esize] = new_port_buf;
+
+  pid_t subchild;
+  sigset_t reset;
+  sigaddset(&reset, SIGCHLD);
+  sigaddset(&reset, SIGINT);
+  sigaddset(&reset, SIGTERM);
+  sigaddset(&reset, SIGPIPE);
+  posix_spawnattr_t attr;
+  posix_spawn_file_actions_t file_actions;
+  posix_spawnattr_init(&attr);
+  posix_spawn_file_actions_init(&file_actions);
+  posix_spawnattr_setflags(&attr, POSIX_SPAWN_CLOEXEC_DEFAULT | POSIX_SPAWN_SETSIGDEF | POSIX_SPAWN_RESETIDS | POSIX_SPAWN_SETPGROUP);
+  posix_spawnattr_setsigdefault(&attr, &reset);
+  posix_spawnattr_setpgroup(&attr, 0);
+  posix_spawn_file_actions_addopen(&file_actions, STDOUT_FILENO, state->stdout, O_WRONLY | O_APPEND | O_CREAT, 0755);
+  posix_spawn_file_actions_addopen(&file_actions, STDERR_FILENO, state->stderr, O_WRONLY | O_APPEND | O_CREAT, 0755);
+  posix_spawn_file_actions_addopen(&file_actions, STDIN_FILENO, "/dev/null", O_RDONLY, 0755);
+
+  int posix_status = posix_spawn(&subchild, "/bin/sh", &file_actions, &attr, argv, new_environ);
+  assert(posix_status == 0);
+  posix_spawn_file_actions_destroy(&file_actions);
+  posix_spawnattr_destroy(&attr);
+
+  fiveman_teardown_sampling(state);
+
+  state->pid = subchild;
+  fiveman_init_sampling(state);
+
+  return subchild;
 }
 
 pid_t fiveman_process_state_signal(fiveman_process_state * state, int signal){
@@ -159,33 +135,7 @@ pid_t fiveman_process_state_signal(fiveman_process_state * state, int signal){
 
 void fiveman_process_state_child_process_status(fiveman_process_state * state) {
   if(fiveman_process_state_is_alive(state)){
-    fd_set set;
-    struct timeval timeout;
-    FD_ZERO (&set);
-    FD_SET (state->host_read_fd, &set);
-    timeout.tv_sec = 0;
-    timeout.tv_usec = 100;
-    const int max_samples_per_read = 1024;
-    const size_t sample_buffer_size = sizeof(fiveman_process_statistics_sample) * max_samples_per_read;
-    fiveman_process_statistics_sample sample_buffer[max_samples_per_read];
-    bzero(sample_buffer, sample_buffer_size);
-    int nread = 0;
-    int nidx = 0;
-    while(1){
-      int res = select(state->host_read_fd + 1, &set, NULL, NULL, &timeout);
-      if(res <= 0){
-        break;
-      }
-      nread = read(state->host_read_fd, sample_buffer, sample_buffer_size);
-      if(nread < 1){
-        break;
-      } else {
-        nidx = (nread / sizeof(fiveman_process_statistics_sample)) - 1;
-        if(nidx >= 0){
-          memcpy(&state->sample, &sample_buffer[nidx], sizeof(fiveman_process_statistics_sample));
-        }
-      }
-    }
+    fiveman_process_state_sample_process(state);
   }
 
 }
@@ -207,30 +157,39 @@ const char * fiveman_get_pager(){
 
 void fiveman_process_state_page_file(fiveman_process_state * state, char * file){
   suspend_screen();
-  pid_t child = fork();
-  reset_signal_handlers();
-  install_ignore_sigint_handler();
-  if(child == 0){ // in child, paging process
-    const char * pager = fiveman_get_pager();
-    execlp(pager, pager, file, NULL);
-    printf("Couldn't page file %s, %i", file, errno);
-    exit(errno);
-  } else if(child > 0){ // in parent
-    // Pass control of the TTY to the child process
-    tcsetpgrp(STDIN_FILENO, child);
-    tcsetpgrp(STDOUT_FILENO, child);
-    tcsetpgrp(STDERR_FILENO, child);
-    int status;
-    waitpid(child, &status, 0);
-    reset_signal_handlers();
-    install_signal_handlers();
-    fiveman_process_state_reap_zombie_processes(state);
-    reap_zombie_processes();
-    resume_screen();
-  } else {
-    assert(child >= 0);
-    exit(-1);
-  }
+
+  const char * pager = fiveman_get_pager();
+  char * argv[] = { (char *)pager, file, NULL };
+
+  pid_t subchild;
+  sigset_t reset;
+  sigaddset(&reset, SIGCHLD);
+  sigaddset(&reset, SIGINT);
+  sigaddset(&reset, SIGTERM);
+  sigaddset(&reset, SIGPIPE);
+  posix_spawnattr_t attr;
+  posix_spawn_file_actions_t file_actions;
+  posix_spawnattr_init(&attr);
+  posix_spawn_file_actions_init(&file_actions);
+  posix_spawnattr_setflags(&attr, POSIX_SPAWN_CLOEXEC_DEFAULT | POSIX_SPAWN_RESETIDS);
+  posix_spawnattr_setsigdefault(&attr, &reset);
+  posix_spawn_file_actions_addinherit_np(&file_actions, STDOUT_FILENO);
+  posix_spawn_file_actions_addinherit_np(&file_actions, STDERR_FILENO);
+  posix_spawn_file_actions_addinherit_np(&file_actions, STDIN_FILENO);
+
+  int posix_status = posix_spawnp(&subchild, pager, &file_actions, &attr, argv, environ);
+  assert(posix_status == 0);
+  posix_spawn_file_actions_destroy(&file_actions);
+  posix_spawnattr_destroy(&attr);
+
+  tcsetpgrp(STDIN_FILENO, subchild);
+  tcsetpgrp(STDOUT_FILENO, subchild);
+  tcsetpgrp(STDERR_FILENO, subchild);
+  int status;
+  waitpid(subchild, &status, WNOHANG);
+  printf("Resuming scren!\n");
+
+  resume_screen();
 }
 
 int fiveman_process_state_lifetime(fiveman_process_state * state){
@@ -266,13 +225,13 @@ void fiveman_process_state_lifetime_str(fiveman_process_state * state, char * bu
   snprintf(buf, buf_size, "%.2f %s", ratio, label);
 }
 
-void fiveman_process_state_page_stdout(fiveman_process_state * state){
-  fiveman_process_state_page_file(state, state->stdout);
+void fiveman_process_state_page_stdout(fiveman_process_state * state, fiveman_pager_fork * fork){
+  fiveman_pager_fork_page_file(fork, state->stdout);
   state->stdout_paged_at = state->stdout_stat.st_mtime;
 }
 
-void fiveman_process_state_page_stderr(fiveman_process_state * state){
-  fiveman_process_state_page_file(state, state->stderr);
+void fiveman_process_state_page_stderr(fiveman_process_state * state, fiveman_pager_fork * fork){
+  fiveman_pager_fork_page_file(fork, state->stderr);
   state->stderr_paged_at = state->stderr_stat.st_mtime;
 }
 
@@ -281,7 +240,11 @@ void fiveman_process_state_change_intent(fiveman_process_state * state, FIVEMAN_
 }
 
 int fiveman_process_state_is_alive(fiveman_process_state * state){
-  return (kill(state->pid, 0) == 0);
+  if(state->pid > 0){
+    return (kill(state->pid, 0) == 0);
+  } else {
+    return 0;
+  }
 }
 
 void fiveman_process_state_converge_start(fiveman_process_state * state, char * directory){
@@ -289,14 +252,12 @@ void fiveman_process_state_converge_start(fiveman_process_state * state, char * 
     if(fiveman_process_state_is_alive(state)){ // Process running
       // Do nothing, already started
     } else {
-      // Start process, had previously stopped and should be started again
-      state->pid = fiveman_process_state_start(state, directory);
+      fiveman_process_state_start(state, directory);
       time(&state->last_state_change);
-
     }
   } else {
     // Start process, never been started before
-    state->pid = fiveman_process_state_start(state, directory);
+    fiveman_process_state_start(state, directory);
     time(&state->last_state_change);
   }
 }
@@ -348,7 +309,7 @@ int fiveman_process_state_status_string(char * buffer, size_t buf_len, int index
 
   if(state->pid > 0){
     pid = state->pid;
-    if(kill(state->pid, 0) == 0){
+    if(fiveman_process_state_is_alive(state)){
       title = ACTIVE_STATUS_TITLE;
     } else {
       title = STOPPED_STATUS_TITLE;
@@ -391,23 +352,12 @@ void fiveman_process_state_converge(fiveman_process_state * state, char * direct
   }
 }
 
-void fiveman_process_state_initialize_pipes(fiveman_process_state * state) {
-  fiveman_process_state_close_pipes(1, 1, state);
-  int fds[2];
-  assert(pipe(fds) == 0);
-  state->host_read_fd = fds[0];
-  state->child_write_fd = fds[1];
+void fiveman_process_state_desires_intent(fiveman_process_state * state, FIVEMAN_INTENT intent){
+  state->desired_intent.intent = intent;
 }
 
-void fiveman_process_state_close_pipes(int close_in, int close_out, fiveman_process_state * state) {
-  if(close_in && state->child_write_fd > -1){
-    close(state->child_write_fd);
-    state->child_write_fd = -1;
-  }
-  if(close_out && state->host_read_fd > -1){
-    close(state->host_read_fd);
-    state->host_read_fd = -1;
-  }
+void fiveman_process_state_reflect_desired_intent(fiveman_process_state * state){
+  state->intent.intent = state->desired_intent.intent;
 }
 
 void fiveman_process_state_reap_zombie_processes(fiveman_process_state * state) {
@@ -417,10 +367,14 @@ void fiveman_process_state_reap_zombie_processes(fiveman_process_state * state) 
 }
 
 void fiveman_process_state_sample_process(fiveman_process_state * state) {
-  fiveman_process_statistics_sample new_sample;
-  bzero(&new_sample, sizeof(fiveman_process_statistics_sample));
-  fiveman_sample_info(&state->sample, &new_sample);
-  fiveman_process_state_set_sample(state, &new_sample);
+  time_t now;
+  time(&now);
+  if( (now - state->sample.sample_time) >= 1){
+    fiveman_process_statistics_sample new_sample;
+    bzero(&new_sample, sizeof(fiveman_process_statistics_sample));
+    fiveman_sample_info(state, &state->sample, &new_sample);
+    fiveman_process_state_set_sample(state, &new_sample);
+  }
 }
 
 void fiveman_process_state_clear_sample(fiveman_process_state * state) {
